Rename mastertalker_api to ot2_workcell_manager_client
