+++ src/arm_controller/package.xml
@@ -14,7 +14,7 @@
 
   <exec_depend>rostalker2interface</exec_depend>
   <exec_depend>arm_client</exec_depend>
-  <exec_depend>mastertalker_api</exec_depend>
+  <exec_depend>ot2_workcell_manager_client</exec_depend>
   <exec_depend>ot2talker_api</exec_depend>
 
   <export>
