+++ src/mastertalker/mastertalker/master.py
@@ -8,7 +8,7 @@ from random import random
 from rostalker2interface.srv import *
 from rostalker2interface.msg import *
 from pathlib import Path
-from mastertalker_api.retry_api import *
+from ot2_workcell_manager_client.retry_api import *
 from mastertalker.worker_thread import worker_class
 
 # Transfer api import
