--- src/mastertalker_api/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/mastertalker_api
-[install]
-install-scripts=$base/lib/mastertalker_api
