+++ src/ot2talker/package.xml
@@ -13,7 +13,7 @@
   <test_depend>python3-pytest</test_depend>
 
   <exec_depend>rostalker2interface</exec_depend>
-  <exec_depend>mastertalker_api</exec_depend>
+  <exec_depend>ot2_workcell_manager_client</exec_depend>
   <exec_depend>arm_client</exec_depend>
   <exec_depend>ot2talker_api</exec_depend>
